requirements.txt
Flask
requests
beautifulsoup4
pandas
openai
jsonify
python-dotenv